requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
pydantic==2.11.4
pydantic-settings==2.9.1
python-multipart==0.0.20
google-generativeai==0.8.5
google-genai==1.33.0
thefuzz[speedup]==0.22.1
